improve azure and aws to use same preprocesing and postprocesing of text like google
